requirements.txt
# Natural Language Processing
langchain==0.0.267
chromadb==0.4.6
pdfminer.six==20221105
InstructorEmbedding
sentence-transformers==2.2.2
faiss-cpu
huggingface_hub==0.25.0
transformers
autoawq; sys_platform != 'darwin'
protobuf==3.20.2; sys_platform != 'darwin'
protobuf==3.20.2; sys_platform == 'darwin' and platform_machine != 'arm64'
protobuf==3.20.3; sys_platform == 'darwin' and platform_machine == 'arm64'
auto-gptq==0.6.0; sys_platform != 'darwin'
docx2txt
unstructured
unstructured[pdf]

# Mac M4 specific dependencies
llama-cpp-python; sys_platform == 'darwin' and platform_machine == 'arm64'
psutil

# NLP toolkit
nltk

# Utilities
urllib3==1.26.6
accelerate
bitsandbytes ; sys_platform != 'win32' and sys_platform != 'darwin'
click
flask
requests

# Streamlit related
streamlit
streamlit-extras

# Excel File Manipulation
openpyxl

# Development and debugging
watchdog